+++ pyproj/crs/crs.py
@@ -1794,6 +1794,23 @@ class CustomConstructorCRS(CRS):
         """
         return CRS(self._crs.to_3d(name=name))
 
+    def to_2d(self, name: str | None = None) -> "CRS":
+        """
+        .. versionadded:: 3.6.0
+
+        Convert the current CRS to the 2D version if it makes sense.
+
+        Parameters
+        ----------
+        name: str, optional
+            CRS name. Defaults to use the name of the original CRS.
+
+        Returns
+        -------
+        CRS
+        """
+        return CRS(self._crs.to_2d(name=name))
+
 
 class GeographicCRS(CustomConstructorCRS):
     """